StableHLO conversion from transpose to permute
- changed `permutation` attribute element type from i32 to i64
- fixed tests

+++ test/ttmlir/Dialect/TTIR/data_movement/permute/permute_tests_negative.mlir
@@ -39,7 +39,7 @@ module {
   func.func @permute_non_valid_permutation(%arg0: tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16> {
     // CHECK: error: 'ttir.permute' op Expected result shape (16, 64, 32), got (16, 32, 64)
     %0 = tensor.empty() : tensor<16x32x64xbf16>
-    %1 = "ttir.permute"(%arg0, %0) <{operand_constraints = [#any_device, #any_device], permutation = array<i32: 0, 2, 1>}> : (tensor<16x32x64xbf16>, tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16>
+    %1 = "ttir.permute"(%arg0, %0) <{operand_constraints = [#any_device, #any_device], permutation = array<i64: 0, 2, 1>}> : (tensor<16x32x64xbf16>, tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16>
     return %1 : tensor<16x32x64xbf16>
   }
 }
